+++ scripts/utils.py
@@ -1,7 +1,9 @@
 import matplotlib.pyplot as plt
 import torch
 import torchvision
-from dataset import EchoDatasetLan
+from dataset_masks import EchoDataset
+
+# from dataset_heatmap import EchoDatasetHeatmap
 from torch.utils.data import DataLoader
 import os
 import cv2
@@ -61,6 +63,7 @@ def get_loaders_masks(
 
     return train_loader, val_loader
 
+
 def get_loaders_landmarks():
     pass
 
@@ -68,10 +71,8 @@ def get_loaders_landmarks():
 def get_loaders_landmarks(
     train_dir,
     train_maskdir,
-    train_heatmaps,
     val_dir,
     val_maskdir,
-    val_heatmaps,
     batch_size,
     train_transform,
     val_transform,
@@ -81,7 +82,7 @@ def get_loaders_landmarks(
     train_ds = EchoDatasetHeatmap(
         image_dir=train_dir,
         mask_dir=train_maskdir,
-        heatmap_dir = train_heatmaps,
+        csv_file="../EchoNet-Dynamic/VolumeTracings.csv",
         transform=train_transform,
     )
 
@@ -96,7 +97,7 @@ def get_loaders_landmarks(
     val_ds = EchoDatasetHeatmap(
         image_dir=val_dir,
         mask_dir=val_maskdir,
-        heatmap_dir = val_heatmaps,
+        csv_file="../EchoNet-Dynamic/VolumeTracings.csv",
         transform=val_transform,
     )
 
@@ -111,7 +112,7 @@ def get_loaders_landmarks(
     return train_loader, val_loader
 
 
-def check_accuracy(loader, model,model_type, device="cuda" ):
+def check_accuracy(loader, model, model_type, device="cuda"):
     num_correct = 0
     num_pixels = 0
     dice_score = 0
@@ -127,9 +128,8 @@ def check_accuracy(loader, model,model_type, device="cuda" ):
                 num_correct += (preds == y).sum()
                 num_pixels += torch.numel(preds)
                 dice = Dice().to(device)
-                preds, y =preds.int(), y.int()
-                dice_score += dice(preds,y)
-
+                preds, y = preds.int(), y.int()
+                dice_score += dice(preds, y)
 
         print(
             f"Got {num_correct}/{num_pixels} with acc {num_correct/num_pixels*100:.2f}"
@@ -139,20 +139,18 @@ def check_accuracy(loader, model,model_type, device="cuda" ):
 
     elif model_type == "landmarks":
         with torch.no_grad():
-            for x,z, y in loader:
-
+            for x, z, y in loader:
                 x = x.to(device)
                 y = y.to(device).unsqueeze(1)
 
                 preds = model(x)
-                #convert heatmap to image
+                # convert heatmap to image
                 preds = heatmap_to_image(preds).to(device)
                 num_correct += (preds == y).sum()
                 num_pixels += torch.numel(preds)
                 dice = Dice().to(device)
-                preds, y =preds.int(), y.int()
-                dice_score += dice(preds,y)
-
+                preds, y = preds.int(), y.int()
+                dice_score += dice(preds, y)
 
         print(
             f"Got {num_correct}/{num_pixels} with acc {num_correct/num_pixels*100:.2f}"
@@ -171,62 +169,65 @@ def folder_creation(folder):
 def save_predictions_as_imgs(
     loader, model, model_type, folder="saved_images/", device="cuda"
 ):
-
     folder_creation(folder)
     model.eval()
     if model_type == "masks":
-
         for idx, (x, y) in enumerate(loader):
             x = x.to(device=device)
             with torch.no_grad():
                 preds = torch.sigmoid(model(x))
                 preds = (preds >= 0.5).float()
-                #save masks predictions as image
+                # save masks predictions as image
                 folder_creation(f"{folder}/masks_predictions")
                 for prediction in range(preds.shape[0]):
-                    torchvision.utils.save_image(preds[prediction], f"{folder}/masks_predictions/batch_{idx}_no{prediction}.png")
+                    torchvision.utils.save_image(
+                        preds[prediction],
+                        f"{folder}/masks_predictions/batch_{idx}_no{prediction}.png",
+                    )
 
-                #Saving original masks as individual images
+                # Saving original masks as individual images
                 folder_creation(f"{folder}/original_masks")
 
                 y[y == 1] = 255
                 for mask in range(y.shape[0]):
-                    torchvision.utils.save_image(y[mask], f"{folder}/original_masks/batch{idx}_no{mask}.png")
+                    torchvision.utils.save_image(
+                        y[mask], f"{folder}/original_masks/batch{idx}_no{mask}.png"
+                    )
 
-                    #cv2.imwrite(f"{folder}/original_masks/batch{idx}_no{mask}.png", y[mask].numpy())
+                    # cv2.imwrite(f"{folder}/original_masks/batch{idx}_no{mask}.png", y[mask].numpy())
 
-                
                 # #Saving the masks as batch
                 # torchvision.utils.save_image(y, f"{folder}/{idx}.png")
-    
 
     elif model_type == "landmarks":
-        for idx, (x,z, y) in enumerate(loader):
+        for idx, (x, z, y) in enumerate(loader):
             x = x.to(device=device)
             with torch.no_grad():
                 preds = torch.sigmoid(model(x))
-        #save heatmaps predictions as images, individually
-        save_batch_to_coordinate(preds,idx, folder)
+        # save heatmaps predictions as images, individually
+        save_batch_to_coordinate(preds, idx, folder)
 
-        #Saving original masks as individual images
+        # Saving original masks as individual images
         folder_creation(f"{folder}/original_masks")
 
         y[y == 1] = 255
         for mask in range(y.shape[0]):
-            torchvision.utils.save_image(y[mask], f"{folder}/original_masks/batch{idx}_no{mask}.png")
-
+            torchvision.utils.save_image(
+                y[mask], f"{folder}/original_masks/batch{idx}_no{mask}.png"
+            )
 
         # #Saving original images as individual images
         # for imagen in range(y.shape[0]):
         #     #cv2.imwrite(f"{folder}/image{idx}_{imagen}.png", x[imagen].permute(1,2,0).to("cpu").numpy())
         #     torchvision.utils.save_image(x[imagen], f"{folder}/image{idx}_{imagen}.png")
 
-        #saving original images as batch
-        #torchvision.utils.save_image(x, f"{folder}/image{idx}.png")
+        # saving original images as batch
+        # torchvision.utils.save_image(x, f"{folder}/image{idx}.png")
 
     model.train()
 
-def unravel_index(indices,shape):
+
+def unravel_index(indices, shape):
     shape = torch.tensor(shape)
     indices = indices % shape.prod()  # prevent out-of-bounds indices
 
@@ -238,42 +239,44 @@ def unravel_index(indices,shape):
 
     return coord.flip(-1)
 
-def save_batch_to_coordinate(t,idx, folder):   
+
+def save_batch_to_coordinate(t, idx, folder):
     for batch in range(t.shape[0]):
         img = np.zeros(t.shape[2:], dtype=np.uint8)
         coor = []
         for i in range(t.shape[1]):
             coordenada = unravel_index(t[batch][i].argmax(), t.shape[2:])
             coor.append(coordenada)
-        points =np.array(coor)
+        points = np.array(coor)
         centroid = np.mean(points, axis=0)
         angles = np.arctan2(points[:, 1] - centroid[1], points[:, 0] - centroid[0])
         sorted_points = points[np.argsort(angles)]
         poly = np.array(sorted_points, np.int32)
         cv2.fillPoly(img, [poly], 255)
 
-        #guardar 
-        folder_creation(f"{folder}/landmark_predictions")
-        cv2.imwrite(f"{folder}/landmark_predictions/prueba_batch{idx}_{batch}.png", img)
+        # guardar
+        cv2.imwrite(f"{folder}/prueba_batch{idx}_{batch}.png", img)
+
 
 def heatmap_to_image(t):
-        masks= torch.tensor([])
-        for batch in range(t.shape[0]):
-            img = np.zeros(t.shape[2:], dtype=np.uint8)
-            coor = []
-            for i in range(t.shape[1]):
-                coordenada = unravel_index(t[batch][i].argmax(), t.shape[2:])
-                coor.append(coordenada)
-            points =np.array(coor)
-            centroid = np.mean(points, axis=0)
-            angles = np.arctan2(points[:, 1] - centroid[1], points[:, 0] - centroid[0])
-            sorted_points = points[np.argsort(angles)]
-            poly = np.array(sorted_points, np.int32)
-            cv2.fillPoly(img, [poly], 255)
+    masks = torch.tensor([])
+    for batch in range(t.shape[0]):
+        img = np.zeros(t.shape[2:], dtype=np.uint8)
+        coor = []
+        for i in range(t.shape[1]):
+            coordenada = unravel_index(t[batch][i].argmax(), t.shape[2:])
+            coor.append(coordenada)
+        points = np.array(coor)
+        centroid = np.mean(points, axis=0)
+        angles = np.arctan2(points[:, 1] - centroid[1], points[:, 0] - centroid[0])
+        sorted_points = points[np.argsort(angles)]
+        poly = np.array(sorted_points, np.int32)
+        cv2.fillPoly(img, [poly], 255)
 
-            mask = torch.from_numpy(img).unsqueeze(0).unsqueeze(0)
-            masks=torch.cat((masks,mask))
-        return masks
+        mask = torch.from_numpy(img).unsqueeze(0).unsqueeze(0)
+        masks = torch.cat((masks, mask))
+    return masks
+
 
 def train_fn(loader, model, optimizer, loss_fn, scaler, DEVICE, model_type):
     loop = tqdm(loader)
@@ -283,14 +286,14 @@ def train_fn(loader, model, optimizer, loss_fn, scaler, DEVICE, model_type):
             data = data.to(device=DEVICE)
             targets = targets.float().to(device=DEVICE)
 
-            #if model is mask, target need color dimension
+            # if model is mask, target need color dimension
             if len(targets.shape) != len(data.shape):
-                targets=targets.unsqueeze(1)
-            
+                targets = targets.unsqueeze(1)
+
             # forward
             with torch.cuda.amp.autocast():
                 predictions = model(data)
-                
+
                 loss = loss_fn(predictions, targets)
             # backward
             optimizer.zero_grad()
@@ -306,14 +309,14 @@ def train_fn(loader, model, optimizer, loss_fn, scaler, DEVICE, model_type):
             data = data.to(device=DEVICE)
             targets = targets.float().to(device=DEVICE)
 
-            #if model is mask, target need color dimension
+            # if model is mask, target need color dimension
             if len(targets.shape) != len(data.shape):
-                targets=targets.unsqueeze(1)
-            
+                targets = targets.unsqueeze(1)
+
             # forward
             with torch.cuda.amp.autocast():
                 predictions = model(data)
-                
+
                 loss = loss_fn(predictions, targets)
             # backward
             optimizer.zero_grad()
